cleaning up rename VM code
